Support typing.Union in FTO077/104
Closes #31
