+++ _flake8_tergeo/checks/ast_subscript.py
@@ -12,6 +12,10 @@ from _flake8_tergeo.ast_util import (
     is_expected_node,
     is_float,
 )
+from _flake8_tergeo.checks.shared import (
+    check_annotation_order,
+    check_bottom_type_in_union,
+)
 from _flake8_tergeo.flake8_types import Issue, IssueGenerator
 from _flake8_tergeo.global_options import get_python_version
 from _flake8_tergeo.registry import register
@@ -23,6 +27,8 @@ def check_subscript(node: ast.Subscript) -> IssueGenerator:
     yield from _check_is_float(node)
     yield from _check_single_item_union(node)
     yield from _check_generator_default(node)
+    yield from _check_annotation_order(node)
+    yield from _check_bottom_type_in_union(node)
 
 
 def _check_is_float(node: ast.Subscript) -> IssueGenerator:
@@ -73,3 +79,19 @@ def _check_generator_default(node: ast.Subscript) -> IssueGenerator:
         issue_number="123",
         message="The 2nd/3rd argument of a Generator annotation can be omitted if they are None.",
     )
+
+
+def _check_annotation_order(node: ast.Subscript) -> IssueGenerator:
+    if not is_expected_node(node.value, "typing", "Union"):
+        return
+    if not isinstance(node.slice, ast.Tuple):
+        return
+    yield from check_annotation_order(node, node.slice.elts)
+
+
+def _check_bottom_type_in_union(node: ast.Subscript) -> IssueGenerator:
+    if not is_expected_node(node.value, "typing", "Union"):
+        return
+    if not isinstance(node.slice, ast.Tuple):
+        return
+    yield from check_bottom_type_in_union(node, node.slice.elts)
